[chore] Remove tests code
